Make HitErrorDisplay an abstract class
